dom/file/tests/test_fileapi_slice.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for File API + Slice</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="common_blob.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>

<body>

<p id="display">
  <canvas id=canvas width=1100 height=1100 hidden moz-opaque></canvas>
</p>

<script type="text/javascript">

SimpleTest.waitForExplicitFinish();

// Create files containing data we'll test with. We'll want long
// strings to ensure they span multiple buffers while loading

let canvasData;
let fileData;

createCanvasURL()
.then(data => {
  canvasData = data;
  return createFile(data, "basicTestFile");
})

.then(file => {
  fileData = file;
  return testSlice(fileData, canvasData.length, "", canvasData, "fileFile");
})

.then(file => {
  let cx = $("canvas").getContext('2d');
  let memFile = cx.canvas.mozGetAsFile("image/png");
  return testSlice(memFile, canvasData.length, "image/png", canvasData, "memFile");
})

.then(SimpleTest.finish);

</script>
</body>
</html>
